[constraint] Split out checker pass 1
